myWebsite/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
MarkupSafe==2.1.3
python-dotenv==1.0.0
markdown==3.4.4
PyYAML==6.0.1
# Pillow==10.0.0  # Commented out due to build issues - only needed for image processing
